+++ example/lib/main.dart
@@ -24,9 +24,26 @@ class _MyAppState extends State<MyApp> {
   @override
   void initState() {
     super.initState();
+
+    SunmiScanner.onScannerStatusChanged().listen((status) {
+      switch (status) {
+        case ScannerConnectionStatus.connected:
+          debugPrint("Connected");
+          break;
+        case ScannerConnectionStatus.disconnected:
+          debugPrint("Disconnected");
+          break;
+        case ScannerConnectionStatus.failedToConnect:
+          debugPrint("Failed To Connect");
+          break;
+      }
+    });
+
     SunmiScanner.onBarcodeScanned().listen((event) {
       _setScannedValue(event);
     });
+
+    SunmiScanner.bindService();
   }
 
   @override